Move existing tests under unit/ folder
